+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+
+## [Unreleased]
+### Added
+- Task priority support with setPriority() method
+
+
 ## [1.2.0] - Current
 ### Added
 - Initial TaskFlow library structure